refactor app name selection to share code between providers
